+++ src/main/java/com/grasshouse/dorandoran/comment/controller/CommentController.java
@@ -1,7 +1,7 @@
 package com.grasshouse.dorandoran.comment.controller;
 
 import com.grasshouse.dorandoran.comment.service.CommentService;
-import com.grasshouse.dorandoran.comment.service.dto.CommentCreateRequest;
+import com.grasshouse.dorandoran.comment.dto.CommentCreateRequest;
 import com.grasshouse.dorandoran.config.jwt.LoginMember;
 import com.grasshouse.dorandoran.member.domain.Member;
 import java.net.URI;
