+++ src/main/java/com/grasshouse/dorandoran/member/controller/MemberController.java
@@ -3,9 +3,9 @@ package com.grasshouse.dorandoran.member.controller;
 import com.grasshouse.dorandoran.config.jwt.LoginMember;
 import com.grasshouse.dorandoran.member.domain.Member;
 import com.grasshouse.dorandoran.member.service.MemberService;
-import com.grasshouse.dorandoran.member.service.dto.MemberResponse;
-import com.grasshouse.dorandoran.member.service.dto.MemberUpdateRequest;
-import com.grasshouse.dorandoran.member.service.dto.MemberUpdateResponse;
+import com.grasshouse.dorandoran.member.dto.MemberResponse;
+import com.grasshouse.dorandoran.member.dto.MemberUpdateRequest;
+import com.grasshouse.dorandoran.member.dto.MemberUpdateResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
