+++ src/main/java/com/grasshouse/dorandoran/member/service/MemberService.java
@@ -2,8 +2,8 @@ package com.grasshouse.dorandoran.member.service;
 
 import com.grasshouse.dorandoran.member.domain.Member;
 import com.grasshouse.dorandoran.member.repository.MemberRepository;
-import com.grasshouse.dorandoran.member.service.dto.MemberUpdateRequest;
-import com.grasshouse.dorandoran.member.service.dto.MemberUpdateResponse;
+import com.grasshouse.dorandoran.member.dto.MemberUpdateRequest;
+import com.grasshouse.dorandoran.member.dto.MemberUpdateResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
