+++ src/main/java/com/grasshouse/dorandoran/post/controller/PostController.java
@@ -3,10 +3,10 @@ package com.grasshouse.dorandoran.post.controller;
 import com.grasshouse.dorandoran.config.jwt.LoginMember;
 import com.grasshouse.dorandoran.member.domain.Member;
 import com.grasshouse.dorandoran.post.service.PostService;
-import com.grasshouse.dorandoran.post.service.dto.PostBoundsRequest;
-import com.grasshouse.dorandoran.post.service.dto.PostCreateRequest;
-import com.grasshouse.dorandoran.post.service.dto.PostCreateResponse;
-import com.grasshouse.dorandoran.post.service.dto.PostResponse;
+import com.grasshouse.dorandoran.post.dto.PostBoundsRequest;
+import com.grasshouse.dorandoran.post.dto.PostCreateRequest;
+import com.grasshouse.dorandoran.post.dto.PostCreateResponse;
+import com.grasshouse.dorandoran.post.dto.PostResponse;
 import java.net.URI;
 import java.util.List;
 import javax.validation.Valid;
