+++ src/main/java/com/grasshouse/dorandoran/post/controller/PostLikeController.java
@@ -3,7 +3,7 @@ package com.grasshouse.dorandoran.post.controller;
 import com.grasshouse.dorandoran.config.jwt.LoginMember;
 import com.grasshouse.dorandoran.member.domain.Member;
 import com.grasshouse.dorandoran.post.service.PostLikeService;
-import com.grasshouse.dorandoran.post.service.dto.PostLikeCreateRequest;
+import com.grasshouse.dorandoran.post.dto.PostLikeCreateRequest;
 import java.net.URI;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
