+++ src/main/java/com/grasshouse/dorandoran/post/service/PostLikeService.java
@@ -11,7 +11,7 @@ import com.grasshouse.dorandoran.post.domain.Post;
 import com.grasshouse.dorandoran.post.domain.PostLike;
 import com.grasshouse.dorandoran.post.repository.PostLikeRepository;
 import com.grasshouse.dorandoran.post.repository.PostRepository;
-import com.grasshouse.dorandoran.post.service.dto.PostLikeCreateRequest;
+import com.grasshouse.dorandoran.post.dto.PostLikeCreateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
