+++ src/main/java/com/grasshouse/dorandoran/post/service/PostService.java
@@ -8,10 +8,10 @@ import com.grasshouse.dorandoran.member.domain.Member;
 import com.grasshouse.dorandoran.post.domain.Post;
 import com.grasshouse.dorandoran.post.repository.PostRepository;
 import com.grasshouse.dorandoran.post.repository.PostRepositorySupport;
-import com.grasshouse.dorandoran.post.service.dto.PostBoundsRequest;
-import com.grasshouse.dorandoran.post.service.dto.PostCreateRequest;
-import com.grasshouse.dorandoran.post.service.dto.PostCreateResponse;
-import com.grasshouse.dorandoran.post.service.dto.PostResponse;
+import com.grasshouse.dorandoran.post.dto.PostBoundsRequest;
+import com.grasshouse.dorandoran.post.dto.PostCreateRequest;
+import com.grasshouse.dorandoran.post.dto.PostCreateResponse;
+import com.grasshouse.dorandoran.post.dto.PostResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
