+++ src/test/java/com/grasshouse/dorandoran/comment/controller/CommentLikeControllerTest.java
@@ -13,7 +13,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.grasshouse.dorandoran.comment.service.CommentLikeService;
-import com.grasshouse.dorandoran.comment.service.dto.CommentLikeCreateRequest;
+import com.grasshouse.dorandoran.comment.dto.CommentLikeCreateRequest;
 import com.grasshouse.dorandoran.common.CommonControllerTest;
 import com.grasshouse.dorandoran.config.jwt.JwtTokenProvider;
 import com.grasshouse.dorandoran.member.repository.MemberRepository;
