+++ src/test/java/com/grasshouse/dorandoran/member/controller/MemberControllerTest.java
@@ -17,8 +17,8 @@ import com.grasshouse.dorandoran.common.CommonControllerTest;
 import com.grasshouse.dorandoran.config.jwt.JwtTokenProvider;
 import com.grasshouse.dorandoran.member.repository.MemberRepository;
 import com.grasshouse.dorandoran.member.service.MemberService;
-import com.grasshouse.dorandoran.member.service.dto.MemberUpdateRequest;
-import com.grasshouse.dorandoran.member.service.dto.MemberUpdateResponse;
+import com.grasshouse.dorandoran.member.dto.MemberUpdateRequest;
+import com.grasshouse.dorandoran.member.dto.MemberUpdateResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
