+++ src/test/java/com/grasshouse/dorandoran/member/service/MemberServiceTest.java
@@ -9,8 +9,8 @@ import com.grasshouse.dorandoran.comment.domain.Comment;
 import com.grasshouse.dorandoran.comment.repository.CommentRepository;
 import com.grasshouse.dorandoran.member.domain.Member;
 import com.grasshouse.dorandoran.member.repository.MemberRepository;
-import com.grasshouse.dorandoran.member.service.dto.MemberUpdateRequest;
-import com.grasshouse.dorandoran.member.service.dto.MemberUpdateResponse;
+import com.grasshouse.dorandoran.member.dto.MemberUpdateRequest;
+import com.grasshouse.dorandoran.member.dto.MemberUpdateResponse;
 import com.grasshouse.dorandoran.post.domain.Post;
 import com.grasshouse.dorandoran.post.repository.PostRepository;
 import org.junit.jupiter.api.AfterEach;
