+++ src/test/java/com/grasshouse/dorandoran/post/controller/PostControllerTest.java
@@ -26,9 +26,9 @@ import com.grasshouse.dorandoran.config.jwt.JwtTokenProvider;
 import com.grasshouse.dorandoran.member.domain.Member;
 import com.grasshouse.dorandoran.member.repository.MemberRepository;
 import com.grasshouse.dorandoran.post.service.PostService;
-import com.grasshouse.dorandoran.post.service.dto.PostCreateRequest;
-import com.grasshouse.dorandoran.post.service.dto.PostCreateResponse;
-import com.grasshouse.dorandoran.post.service.dto.PostResponse;
+import com.grasshouse.dorandoran.post.dto.PostCreateRequest;
+import com.grasshouse.dorandoran.post.dto.PostCreateResponse;
+import com.grasshouse.dorandoran.post.dto.PostResponse;
 import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
