+++ src/test/java/com/grasshouse/dorandoran/post/controller/PostFilterControllerTest.java
@@ -10,7 +10,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.grasshouse.dorandoran.common.CommonControllerTest;
 import com.grasshouse.dorandoran.post.service.PostFilterService;
-import com.grasshouse.dorandoran.post.service.dto.PostResponse;
+import com.grasshouse.dorandoran.post.dto.PostResponse;
 import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
