+++ src/test/java/com/grasshouse/dorandoran/post/service/PostLikeServiceTest.java
@@ -14,7 +14,7 @@ import com.grasshouse.dorandoran.post.domain.PostLike;
 import com.grasshouse.dorandoran.post.repository.PostLikeRepository;
 import com.grasshouse.dorandoran.post.repository.PostRepository;
 import com.grasshouse.dorandoran.post.repository.PostRepositorySupport;
-import com.grasshouse.dorandoran.post.service.dto.PostLikeCreateRequest;
+import com.grasshouse.dorandoran.post.dto.PostLikeCreateRequest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
